Bugfix: UserGuitar now points to correct guitar_id
UserGuitar's controller create action was missing a `where` clause,
where the params were passed for the query to find in order for
first_or_initialize to work correctly.

UserGuitar is now pointing to correct guitar if it exists, or creating
that guitar then pointing to it if it's a new entry.

TODO: add pg_search fuzzy matching functionality to smooth out typos and
other small differentiations.