Reorder model has_many, through relationships
Turns out the top-to-bottom order of model relationships has an impact
on those relationships. A user trying to 'get' guitars couldn't do so
when has_many guitars, through: user_guitars appeared before has_many
user_guitars.

Not that I necessarily want users to be able to 'get' guitars, but
that's a different decision. Fixing the
ActiveRecord::HasManyThroughOrderError took priority.