Make userguitar create conditional on guitar.save
Now UserGuitar.build only runs if a guitar is successfully saved first.
Perhaps having the guitar creation fail if the userguitar creation fails
is a better idea, though I'm not sure that's necessarily true. I'm fine
with users creating a guitar model even if there was a problem with the
userguitar. Just means the next time they try to create the userguitar,
the guitar model they created previously is in the database. If it was
created successfully I see no reason it shouldn't be saved.