Create working avg_price method
Changed route (and script to match) to /avg_price/:id since Rails was
interpreting /user_guitars/avg_price as 'avg_price' was the ID of the
user_guitar, which is clearly wrong.

Controller action checks that avg_price is a BigDecimal since that's
what SQL saves it as, and returns it if so.